Integrating class into component
